Correctly report both X10 and Insteon group addresses. Reimplmented hashCode and equals using eclipse's Source->Generate hashcode() and equals()...

+++ bundles/binding/org.openhab.binding.insteonplm/src/main/java/org/openhab/binding/insteonplm/internal/device/InsteonDevice.java
@@ -250,7 +250,7 @@ public class InsteonDevice {
 		if (group != -1) {
 			flags |= 0xc0; // mark message as group message
 			// and stash the group number into the address
-			addr = new InsteonAddress(0,0,(byte)(group & 0xff));
+			addr = new InsteonAddress((byte) 0, (byte) 0, (byte) (group & 0xff));
 		} else {
 			addr = getAddress();
 		}
